App2Clap: Better handling for larger host buffer sizes; e.g. 2048 samples.
Windows will only buffer 3 packets at a time.
After that, it just drops packets.
If the host max frame count is larger than that, capture audio in a background thread to avoid continual buffer underruns.
Note that the thread is less optimal (and results in glitches) when the host max frame count is lower, so we don't use a thread for smaller host max frame counts.