Make it possible to click GUI controls with the mouse. Fixes #2.
In order for the tab key to work, we have a hack which uses the grandparent HWND instead of the parent wrapper HWND that REAPER wants us to use.
Previously, the wrapper HWND was the target for hit testing (including mouse clicks), since it's higher in the z-order.
To fix this, hide the wrapper HWND.